Use Action based retry mechanism
