+++ pulsar-functions/utils/src/test/java/org/apache/pulsar/functions/utils/ActionsTest.java
@@ -16,8 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.functions.runtime;
+package org.apache.pulsar.functions.utils;
 
+import org.apache.pulsar.functions.utils.Actions;
 import org.testng.annotations.Test;
 
 import java.util.function.Supplier;
@@ -29,23 +30,23 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.testng.Assert.assertEquals;
 
-public class RuntimeUtilsTest {
+public class ActionsTest {
 
     @Test
     public void testActions() throws InterruptedException {
 
         // Test for success
-        Supplier<RuntimeUtils.Actions.ActionResult> supplier1 = mock(Supplier.class);
-        when(supplier1.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(true).build());
+        Supplier<Actions.ActionResult> supplier1 = mock(Supplier.class);
+        when(supplier1.get()).thenReturn(Actions.ActionResult.builder().success(true).build());
 
-        Supplier<RuntimeUtils.Actions.ActionResult> supplier2 = mock(Supplier.class);
-        when(supplier2.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(true).build());
+        Supplier<Actions.ActionResult> supplier2 = mock(Supplier.class);
+        when(supplier2.get()).thenReturn(Actions.ActionResult.builder().success(true).build());
 
         Runnable onFail = mock(Runnable.class);
         Runnable onSucess = mock(Runnable.class);
 
-        RuntimeUtils.Actions.Action action1 = spy(
-                RuntimeUtils.Actions.Action.builder()
+        Actions.Action action1 = spy(
+                Actions.Action.builder()
                         .actionName("action1")
                         .numRetries(10)
                         .sleepBetweenInvocationsMs(100)
@@ -55,15 +56,15 @@ public class RuntimeUtilsTest {
                         .onSuccess(onSucess)
                         .build());
 
-        RuntimeUtils.Actions.Action action2 = spy(
-                RuntimeUtils.Actions.Action.builder()
+        Actions.Action action2 = spy(
+                Actions.Action.builder()
                         .actionName("action2")
                         .numRetries(20)
                         .sleepBetweenInvocationsMs(200)
                         .supplier(supplier2)
                         .build());
 
-        RuntimeUtils.Actions actions = RuntimeUtils.Actions.newBuilder()
+        Actions actions = Actions.newBuilder()
                 .addAction(action1)
                 .addAction(action2);
         actions.run();
@@ -77,16 +78,16 @@ public class RuntimeUtilsTest {
         // test only run 1 action
 
         supplier1 = mock(Supplier.class);
-        when(supplier1.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(true).build());
+        when(supplier1.get()).thenReturn(Actions.ActionResult.builder().success(true).build());
 
         supplier2 = mock(Supplier.class);
-        when(supplier2.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(true).build());
+        when(supplier2.get()).thenReturn(Actions.ActionResult.builder().success(true).build());
 
         onFail = mock(Runnable.class);
         onSucess = mock(Runnable.class);
 
         action1 = spy(
-                RuntimeUtils.Actions.Action.builder()
+                Actions.Action.builder()
                         .actionName("action1")
                         .numRetries(10)
                         .sleepBetweenInvocationsMs(100)
@@ -97,7 +98,7 @@ public class RuntimeUtilsTest {
                         .build());
 
         action2 = spy(
-                RuntimeUtils.Actions.Action.builder()
+                Actions.Action.builder()
                         .actionName("action2")
                         .numRetries(20)
                         .sleepBetweenInvocationsMs(200)
@@ -106,7 +107,7 @@ public class RuntimeUtilsTest {
                         .onSuccess(onSucess)
                         .build());
 
-        actions = RuntimeUtils.Actions.newBuilder()
+        actions = Actions.newBuilder()
                 .addAction(action1)
                 .addAction(action2);
         actions.run();
@@ -120,16 +121,16 @@ public class RuntimeUtilsTest {
         // test retry
 
         supplier1 = mock(Supplier.class);
-        when(supplier1.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(false).build());
+        when(supplier1.get()).thenReturn(Actions.ActionResult.builder().success(false).build());
 
         supplier2 = mock(Supplier.class);
-        when(supplier2.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(true).build());
+        when(supplier2.get()).thenReturn(Actions.ActionResult.builder().success(true).build());
 
         onFail = mock(Runnable.class);
         onSucess = mock(Runnable.class);
 
         action1 = spy(
-                RuntimeUtils.Actions.Action.builder()
+                Actions.Action.builder()
                         .actionName("action1")
                         .numRetries(10)
                         .sleepBetweenInvocationsMs(10)
@@ -140,14 +141,14 @@ public class RuntimeUtilsTest {
                         .build());
 
         action2 = spy(
-                RuntimeUtils.Actions.Action.builder()
+                Actions.Action.builder()
                         .actionName("action2")
                         .numRetries(20)
                         .sleepBetweenInvocationsMs(200)
                         .supplier(supplier2)
                         .build());
 
-        actions = RuntimeUtils.Actions.newBuilder()
+        actions = Actions.newBuilder()
                 .addAction(action1)
                 .addAction(action2);
         actions.run();
